+++ AverageColorLib/AverageColor_Task.h
@@ -5,7 +5,7 @@
 
 // Chunksize is the approximate size of one chunk of work for a task, in the number of
 // pixels to process. If 0, the number is set to a reasonable value.
-unsigned long AverageColor_Task(
+DWORD AverageColor_Task(
         RawBitmap::PixelColorVector::iterator begin,
         RawBitmap::PixelColorVector::iterator end,
-        unsigned int chunkSize);
+        UINT chunkSize);
